update code and logic

+++ src/main/java/com/linh/warehouse/repository/SaleInvoicePaymentRepository.java
@@ -6,11 +6,12 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface SaleInvoicePaymentRepository extends JpaRepository<SaleInvoicePayment, Integer> {
     List<SaleInvoicePayment> findBySaleInvoiceId(int invoiceId);
     @Query("SELECT SUM(p.amount) FROM SaleInvoicePayment p WHERE p.saleInvoice.id = :invoiceId")
-    Double getTotalPaidAmountByInvoiceId(@Param("invoiceId") int invoiceId);
+    BigDecimal getTotalPaidAmountByInvoiceId(@Param("invoiceId") int invoiceId);
 }
